Conflicts:
	IslandFurnitureERPSystem/IslandFurnitureERPSystem-ejb/src/java/SessionBean/SCM/PurchaseOrderManagementModule.java

+++ IslandFurnitureERPSystem/IslandFurnitureERPSystem-ejb/src/java/Entity/Factory/FactoryRawMaterialAmountEntity.java
@@ -5,7 +5,6 @@
  */
 package Entity.Factory;
 
-import Entity.Factory.MRP.PlannedOrderEntity;
 import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -27,9 +26,9 @@ public class FactoryRawMaterialAmountEntity implements Serializable {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long rawMaterialAmountId;
     private String unit;
-    private Integer amount;
+    private Double amount;
     
-    //raw material amount entity -- factory raw material entity: M <--> 1
+    //raw material amount entity -- factory raw material entity: M --> 1
     @ManyToOne
     private FactoryRawMaterialEntity factoryRawMaterial;
     
@@ -41,11 +40,11 @@ public class FactoryRawMaterialAmountEntity implements Serializable {
         this.unit = unit;
     }
 
-    public Integer getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
